feat: Q-Table implemented and can pass one pipe
There are too many states for the bird to learn to pass multiple pipe
levels. The rewards are too inconsistent if you set the score-limit to
more than 1. After setting only 1 pipe height, and the score-limit=1
and adding eligibility traces. The bird correctly understands the
mapping between states and actions. All data is saved inside the q-table.