feat: incorporated eval_policy into checkpoint
During training, the policy is evaluated after a few timesteps.
It works by using a chain of two callbacks.

The learn function uses the Every N Timesteps callback, which invokes
a evaluate_policy fn after the specified number of timesteps.

The evaluate_policy fn invokes the Scorecallback which fetches the score
after each episode and populates it in the private scores attribute.